service/wiki-api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic-settings>=2.0.0
PyYAML>=6.0.1
openai>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
